more refactoring of Language Banner
- Rename i18nBanner to LanguageSelectBanner
- Delete LanguageWrapper and move needed functionality into App.js
- update Cookie use in banner
