+++ src/css/coa.scss
@@ -28,6 +28,6 @@
 @import 'css/page_sections/main';
 @import "css/page_sections/Header";
 @import "css/page_sections/Footer";
-@import "css/page_sections/I18nBanner";
+@import "css/page_sections/LanguageSelectBanner";
 @import "css/page_sections/Navmenu";
 @import 'css/page_sections/ApplicationBlock';
