+++ src/js/page_sections/LanguageSelectBanner.js
@@ -1,12 +1,13 @@
 import React, { Component } from 'react';
 import { Link } from 'react-static';
+import Cookies from 'js-cookie'
 
 import { SUPPORTED_LANGUAGES, SUPPORTED_LANG_CODES } from 'js/constants/languages'
 import CaretDownSVG from 'js/svg/CaretDown';
 import CaretUpSVG from 'js/svg/CaretUp';
 import { getPathnameWithoutLangCode } from 'js/helpers/language';
 
-class I18nBanner extends Component {
+class LanguageSelectBanner extends Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -15,12 +16,12 @@ class I18nBanner extends Component {
   }
 
   getBannerClassName = () => {
-    const baseClass = 'coa-I18nBanner';
+    const baseClass = 'coa-LanguageSelectBanner';
     return `${baseClass} ${this.state.isOpen ? `${baseClass}--open` : ''}`;
   }
 
   getLanguageItemClassName = (language) => {
-    const blockElementClassname = `coa-I18nBanner__language-item`;
+    const blockElementClassname = `coa-LanguageSelectBanner__language-item`;
     let classNames = blockElementClassname;
 
     if (language.code === this.props.match.params.lang) {
@@ -31,13 +32,14 @@ class I18nBanner extends Component {
   }
 
   getMenuClassName = () => {
-    const base = `coa-I18nBanner__dropdown-menu`;
+    const base = `coa-LanguageSelectBanner__dropdown-menu`;
 
     return this.state.isOpen ? `${base} ${base}--open` : `${base} ${base}--closed`;
   }
 
   handleSetLanguage = (e) => {
     if (e.target.lang) {
+      this.props.updateLanguage(e.target.lang)
       this.setState({
         isOpen: false,
       });
@@ -51,8 +53,9 @@ class I18nBanner extends Component {
   }
 
   getActiveLanguageTitle = () => {
+    const setLanguage = Cookies.get('lang') || this.props.match.params.lang;
     let activeLanguage = SUPPORTED_LANGUAGES.find((lang) => {
-      return lang.code === this.props.match.params.lang;
+      return lang.code === setLanguage;
     });
     activeLanguage = activeLanguage || SUPPORTED_LANGUAGES[0];
     return activeLanguage.title;
@@ -72,11 +75,11 @@ class I18nBanner extends Component {
         <div className="wrapper">
           <div className="row">
             <div className="col-md-12 col-xs-12">
-              <div className="coa-I18nBanner__choose-language"
+              <div className="coa-LanguageSelectBanner__choose-language"
                 onClick={this.handleExpandMenu}
               >
                 <span>Choose Language
-                  <span className="coa-link coa-I18nBanner__choose-language-link">
+                  <span className="coa-link coa-LanguageSelectBanner__choose-language-link">
                     {languageTitle}
                   </span>
                 </span>
@@ -87,10 +90,10 @@ class I18nBanner extends Component {
                 }
                 <div className={this.getMenuClassName()}>
                   <div className="wrapper">
-                    <h4 className="coa-I18nBanner__menu-header">
+                    <h4 className="coa-LanguageSelectBanner__menu-header">
                       Translated by City of Austin
                     </h4>
-                    <ul className="coa-I18nBanner__language-list">
+                    <ul className="coa-LanguageSelectBanner__language-list">
                       { SUPPORTED_LANGUAGES.map((language, i) => {
                         return (
                           <li onClick={this.handleSetLanguage}
@@ -109,7 +112,7 @@ class I18nBanner extends Component {
                     }
                   </ul>
                 </div>
-                <p className="coa-link coa-I18nBanner__cancel"
+                <p className="coa-link coa-LanguageSelectBanner__cancel"
                   onClick={this.handleCancel}
                 >
                   Cancel
@@ -125,4 +128,4 @@ class I18nBanner extends Component {
 
 }
 
-export default I18nBanner;
+export default LanguageSelectBanner;
